(develop) Corrects fixture support
https://app.asana.com/0/1132578874282309/1132590387550875
- Changes support for fixtures for setting up the local dev environment.

Notes: Neither of the drupal fixture modules seem to work out of the box.
I'm having a devil of a time getting this working.  Committing now because
this configuration has no errors, and at least gets the right modules enabled
out of the gate in the dev environment, and corrects the admin password.

Usage:
ddev stop --remove-data --omit-snapshot && ddev start